requirements.txt
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
selenium==3.141.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.1
trio-websocket==0.10.3
urllib3==1.26.15
wsproto==1.2.0
